moved the Header tests to header_test.go
