+++ handler_test.go
@@ -240,11 +240,3 @@ func TestDirSplit(t *testing.T) {
 		})
 	}
 }
-
-func TestHeader(t *testing.T) {
-	ts := httptest.NewServer(Header("superheader", "secret value", Success("yay")))
-
-	resp, err := http.Get(ts.URL)
-	assert.Nil(t, err)
-	assert.Equal(t, "secret value", resp.Header.Get("superheader"))
-}
